CacheStorage: Ignore VARY headers on opaque responses.
The CacheStorage spec deals with VARY headers in two places.  In both
of these places the spec uses the headers exposed by the filtered
response and does not "pierce" the filtering to look at headers inside
an opaque response.

1. Step 7.2.1 of https://w3c.github.io/ServiceWorker/#cache-put uses the
   "inner response" from the "Response" binding object.  This is just
   the non-exposed "little r" response.  It is not the filtered
   response's "internal response".  Later put() stores a clone of this
   response in storage.
2. Step 7 of https://w3c.github.io/ServiceWorker/#request-matches-cached-item-algorithm
   then just reads the exposed headers from the stored response.  Again,
   it does not pierce the filtered response to get the internal response
   and its headers.

This means that opaque responses should act as if they have no VARY
header.  Chrome was already doing this for (1), but not (2).  This CL
fixes this inconsistency.  It also adds WPT tests for these scenarios.

Bug: 1274098
Change-Id: I97091451cc1205f4a3cafe8f2109245d70e3a119
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3307784
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947009}
NOKEYCHECK=True
GitOrigin-RevId: 63f5b5eda46973a567de9f640f9602b82daa1535