Streamline websocket interruption logic and fix issue when reading frame is not restarted after the interceptor interruption.
Replace various flags and checks with two 'interruption_reason' members
and 'resume_reason' parameter.

I believe this is needed as my previous attempt to fix this logic
(crrev.com/c/3284985) was not enough.

For reference, here's a pseudo-UML describing the current code at HEAD:
https://docs.google.com/drawings/d/1GangzHREXPyJg6vGoZxpXSg3BzSw0oGZz-5PDnlbVFE/edit?usp=sharing

Conceptually all these wait_for_* and IsFrameStarted are
equivalent to `await`, i.e. we want to resume the function on the next
line once the async condition is fulfilled. This code is trying to make
a step in this directions while keep the delta as small as possible.

Bug: 423246
Change-Id: Idaabdbc57db97cd9a27c38aedf114eaf4fa3da35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3298020
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947015}
NOKEYCHECK=True
GitOrigin-RevId: 4ab081e31f7714ee07f685aec2781aa4285e9804