[FlexNG] Update the total intrinsic block-size after break before
As a follow up to CL:3242462, update the total intrinsic block-size
when we expand as a result of a break before. To make this easier,
this CL also pulls the total intrinsic block-size into a member
variable.

This update shouldn't result in any functional change at the moment
but may come into play when handling fragmentation for flex rows.

Bug: 660611
Change-Id: If5004da2bccef755200fc296f3dc5e9e87f405c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3308853
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#947028}
NOKEYCHECK=True
GitOrigin-RevId: dc96b40a92aff83dc9c730017590fbc77caf5127