Draw coplanar planes in paint order, ignoring whether front/back facing.
This draws coplanar planes in paint order rather than considering
whether they are front or back facing.  This removes a bit of code that
was intentionally written to consider which way the planes are facing.

This change matches what is required by
https://drafts.csswg.org/css-transforms-2/#3d-rendering-contexts
("Coplanar 3D transformed elements are rendered in painting order") and
matches other browsers.  The behavior also makes sense to me since it
means that 3D-transformed planes that are coplanar more closely match
the behavior of 2D-transformed or untransformed elements, which also
follow the standard paint order.  This also means that changes to what
gets composited will no longer trigger paint order changes due to this
code; previously a change to what is compositing could change which
objects are handled by this sort, which can in turn change the paint
order when those things are coplanar (which they presumably are if they
might not have been composited).

Note that this changes the sense of which coplanar planes are in "front"
and which are in "back", because I believe this was previously reversed.

On the fixed test, this fixes Chromium to match other browser engines.

Fixed: 1261895
Change-Id: I5bddaa6d40cd4f2d368477e9f39e8c78fb22abf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3251501
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947042}
NOKEYCHECK=True
GitOrigin-RevId: 07fe9e9dd89fca8d512df87a29cfdc83a25d1418