[FlexNG] Handle NGBreakStatus::kDisableFragmentation
Add support for handling NGBreakStatus::kDisableFragmentation. In order
to get this to work, we also needed to update a few instances that
checked for HasBlockFragmentation() with InvolvedInBlockFragmentation().
The reason being that if we are relaying out with fragmentation
disabled, we still want to go down the fragmentation specific layout
path so that we don't relayout all items.

Bug: 660611
Change-Id: Id6eaf8ae1e027dfd7e86c6372be1d0eef3949d6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3309023
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947076}
NOKEYCHECK=True
GitOrigin-RevId: 6b9f157b0558e0dc7d6a35891edbb27a0e813e96