Prevent GC during Fast API Call in HTMLCanvasElement::UpdateMemoryUsage
This change fixes a violation of the 'IsAllocationAllowed' constraint
that happened when a canvas backing buffer is lazily reallocated during
a V8 fast API call.  The allocation of the backing does not violate the
constraint because no garbage collected objects are involved.  However,
reporting the allocation to V8 (as externally allocated memory) was
problematic because it could trigger garbage collection.  The solution
implemented in this CL is to use a DeferrablAction to report the
allocation to V8.  This way, the garbage collection, if any, is delayed
until the Fast API call scope has exited.

For regression testing, it is very hard to write reliable tests that
depend on a GC being triggerer  Instead, a conservative DCHECK was
added to ensure that we do no report external allocations to V8 while
inside the scope of a fast API call.  Without the fix in this CL,
several unit tests and web_tests would trip that DCHECK.

BUG=1275084

Change-Id: I0e00d7ec85bd98b7dec772add36ae1cd390d49dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3306809
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947138}
NOKEYCHECK=True
GitOrigin-RevId: 2f389f71d15c58e4a575b76849817554b9e71127