[SelectMenu] Reflect option.selected for selectmenu.value
This CL adds support so that updating option.selected will also update
selectmenu.value.

When the selected value is true, the option is marked as selectmenu's
selectedOption. For now, when the selected value is changed to false,
the first option is going to be selected (support for reset to default
selection is going to be added in a follow up CL).

Bug: 1121840
Change-Id: I7f4b9de685cefeb2f549a3e39b1c69f555490d90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3287417
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#947707}
NOKEYCHECK=True
GitOrigin-RevId: a81b9cafd3147b05a25878da14106e7f8b04a8d9