Unify template to create canvas and offscreencanvas tests
The typical way to add tests for canvas and offscreencanvas is as
following:
- write the test in yaml/element (create the test for canvas)
- copy the the test to yaml/offscreencanvas
- add t.done() to the end of test (create the test for offscreencanvas)

This is because the API for canvas and offscreencanvas are mostly the
same.

This cl updates templates and gentestutils to allow the framework of
generate canvas and offscreencanvas test from test.

I removed both yaml/element/color_space.yaml and
yaml/offscreencanvas/color_space.yaml and add yaml_new/color_space.yaml.
Now the yaml_new/color_space.yaml generates the same results as before.

Note that it's possible to add test for a particular canvas type
by using canvasType as keyword.

Bug: 1275750
Change-Id: I6bf48ec3081f5fe3669bf585f92b39a712045c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3307080
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947713}
NOKEYCHECK=True
GitOrigin-RevId: b9b6f61df6033c568d6a75ecb0769f9ab557ab3c