[GridFragmentation] Basic support of break-before/break-after on item.
This implements basic support for break-before/break-after. At the
moment it only support the top level break-before/break-after value (not
the nested values, until we set up the non-fragmented constraint space
correctly).

The break-before/break-after values are collected at the grid row level
when we calculate the grid (stored in a vector).

Currently this also doesn't (yet) search for the row with the highest
appeal.

Bug: 614667
Change-Id: I0299fa5d7b04106641903607c84fe45470f8a27c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3287251
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#947733}
NOKEYCHECK=True
GitOrigin-RevId: 97724fa3592d935d81aae0d8a0bf0d926c1f46e0