If a block fits, put some content inside before breaking.
If a block fits in the current fragmentainer, don't let overflowing
children pull it over to the next fragmentainer because of less
appealing breaks inside. Instead overflow the fragmentainer.

This is a special rule that is required in order to print Google Docs
correctly (maybe there are other reasons too). Gecko also does this.
EdgeHTML also does something similar. See crbug.com/507972

The legacy block fragmentation engine also had a workaround for Google
Docs, although simpler and weirder (if a line doesn't fit in the current
fragmentainer, but won't fit in the next one either, just leave it where
it is, even if we had already made some progress in the current
fragmentainer).

In a way, crbug.com/178742#c8 is related, which is about overflow inside
monolithic content, and how we let it overflow the fragmentainer.

The logic in IsNodeFullyGrown() was very similar to what we needed here,
so remove it and share code and logic.

Bug: 1242348
Change-Id: Ibc26576d3096f59b629e1ef8c5f725621d8094e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3310887
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947906}
NOKEYCHECK=True
GitOrigin-RevId: 6b75c95b22ff3842fb3fa4849dfaf18a74b063c0