[web snapshots] Prototype Blink integration for web snapshots
This solution is based on recognizing magic bytes at the beginning of
the script resource.

The flag is only suitable for local testing, not suitable for origin
trials. This is expected to be removed (and the real Blink integration
landed) before M102.

Bug: v8:11525,chromium:1173534
Change-Id: I506e649393e430da29f6ac13a9a896f0e2a06e53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3301046
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947947}
NOKEYCHECK=True
GitOrigin-RevId: ac89091e0e0761c5fa5a70fb8cb81b540d9675c2