Add FlexOffset
Replace the use of LayoutPoint for flex item offsets in the various
flex algorithms with a new FlexOffset type to make it clearer which is
the main/cross axis offset. Additionally, when using this offset in
NG, convert it to LogicalOffset to make it clear which is the
block/inline offset.

This CL should not result in any functional changes.

Bug: 660611
Change-Id: Ic04caa38e7a6b0d09685e42763fce8f764a18590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3313425
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#948030}
NOKEYCHECK=True
GitOrigin-RevId: 975566628c7a51a8b3a6b2adfe8f7dceef6b5db5