[MPArch] Disallow use of WebBluetoothServiceImpl with fenced-frames
This CL disallows the use of WebBluetoothServiceImpl when dealing with a
render frame host that is fenced frame, or a sub frame where one of its
parent frames is a fenced frame. With these changes, instances of
WebBluetoothServiceImpl are prevented to be created in the first place,
so no additional checks for fenced frames are required in the instance
after creation, since the main checks are done in the renderer.

Additionally, this CL is also correcting the occurrences of GetMainFrame
to a more appropriate access.

Bug: 1275606
Change-Id: I770ce79c726977c807f0ac7d6bbce0da4fa5e461
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3310879
Commit-Queue: Claudio DeSouza <cdesouza@igalia.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948175}
NOKEYCHECK=True
GitOrigin-RevId: de4a42dd19de7d004756dfd48256898473a712c4