adpf: Hook up viz clients threads
This CL adds *some* threads in browser and renderer that are on the
critical path for generating frames to the performance hint session in
the viz/GPU process.

The specific threads are:
* renderer compositor thread
* renderer IO thread
* video submitter thread
* browser UI (ie compositor) thread
Note browser IO thread is left out due to complexity of getting the
thread ID though it should really be included. These threads are
generally enough for example to include all threads involved in a
compositor touch scroll.

And intentionally excluded are raster threads and the renderer main
thread. Here are some reasons for excluding these threads for now:
* These threads are not involved in all frames, and can often run behind
  viz frames. Eg if renderer main thread is slow, it can run at a lower
  frame rate than viz. So it may not make sense to add it to the viz
  frame.
* It's expensive to adjust the threads in the performance hint session,
  and it's possible these threads are added / removed on a per-frame
  basis. And the OS can not react quickly enough to threads changing in
  every frame for it to make sense.
* It's a lot of complexity to keep track of which threads are actually
  on the critical path of a particular frame.

Rough implementation description:

Add SetThreadIds to CompositorFrameSink mojom. Then have each
viz::Surface keep track of the client threads. At draw time after
aggregation, also aggregate the threads from among all the Surfaces that
are in the current frame, and add those to the hint session. Note it's
technically makes more sense to include Surfaces that has new frame
since last draw; but that gets into the problem potentially changing
threads every frame. There's quite a bit of plumbing to pass the IO
thread ID to AsyncLayerTreeFrameSink.

Security:

Renderer processes on Android are not directly children of browser
process and they are under different UIDs. So browser (or GPU) does not
have any permission to inspect the process to check thread IDs reported
by a renderer process belongs to that renderer process.

The OS will check that all thread IDs reported by chromium belongs to
chromium. So the best we can do is to ensure that the thread IDs
reported by renderers do not belong to the browser or GPU processes, the
only two privileged processes on Android. This is done by passing the
browser's PID to the GPU, and then GPU process can perform the check for
both browser and GPU process by reading /proc. This is all
Android-specific, so the mojom changes are added for Android only as
well; though the rest of the code is added cross-platform where
possible.

Bug: 1157620
Change-Id: Ibf6282a3f74168c246c7c32f5d2ecd92554e30dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3156346
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948794}
NOKEYCHECK=True
GitOrigin-RevId: e0cdbd9e151bd099f29e7064604f4176efcf1472