WebGPU: Fix setBindGroup argument optionality
setBindGroup was previously split into additional overloads to avoid
allocation of an empty vector for the default dynamicOffsets argument.
This incorrectly made it impossible to pass |undefined| as the third
argument since it was not optional. This fixes the idl to make the arg
optional while preserving the original intent to avoid allocation of an
empty vector. C++ overloads without the vector argument are still called
if no third arg exists, or if undefined is passed.

Bug: 852089
Change-Id: I29cd291126d7efbd3586bd61e6454bf1fc7eff0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3318930
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948844}
NOKEYCHECK=True
GitOrigin-RevId: 988b623fdf134dbc5aa69311a7e2b7dc376a34b4