Fenced Frames: Enforce opt-in via S-L-M for nested iframes as well
This CL implements the enforced opt-in via supports-loading-mode for
nested iframes inside a fenced frame. The "Sec-Fetch-Dest" header value
will be "fencedframe" for the nested iframes as well to signal to the
servers that they need to opt-in in order to load.

Bug: 1276425
Change-Id: I5cc59806701e63b21d217e320ca349f615d47c76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3313886
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948982}
NOKEYCHECK=True
GitOrigin-RevId: 26c4857b6035265e3eb074a052794fbfd9f7415c