Unconditionally initialize the AppHistory object during commit.
If AppHistory is ever going to be used, it must be initialized during
commit, because it requires state that is only available during commit.
However, when AppHistory is enabled by an origin trial <meta> tag,
AppHistory will not be enabled until after commit, and AppHistory
has inconsistent state because the commit-time initialization was
skipped. Therefore, do the commit-time initialization even if
AppHistory is disabled, in case the origin trial is later enabled by
a <meta> tag. This means that some AppHistory state will be
initialized and stored even if AppHistory is disabled, but it won't
be web-visible, no events will fire, etc.

Bug: 1276587
Test: http/tests/origin_trials/webexposed/app-history-origin-trial.html
Change-Id: Ibb40d91d9b0dc20bdf7288b3648055b563ad407d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3314605
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Auto-Submit: Nate Chapin <japhet@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949112}
NOKEYCHECK=True
GitOrigin-RevId: 2a5d607f5304cd85fa35bbb6dff791c1aeeece48