Do not display grammar error if it overlaps with spellcheck error
This is a new implementation to resolve grammar check and spellcheck
conflict, the old implementation in crrev.com/c/3159841 was reverted
because of a performance issue. I have verified this new implementation
won't cause that issue any more. The only drawback is that the IME side
won't be able to know if a grammar check marker is added or not, so the
InputMethod.Assistive.Grammar.Actions.Underlined metric may not be very
accurate.

Bug: 1249351
TEST: added unit test and also verified on DUT.
Change-Id: If622d20aa69c59a4d6b6d9f70d0a2c6b65c338d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3317002
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Jing Wang <jiwan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949288}
NOKEYCHECK=True
GitOrigin-RevId: e9a10918149a5773c1aba0c45c1dc07d62fd50a1