Notify of the frame color scheme has been inferred.
This patch infers the document's color scheme according to the
background color, this approach assumes that the background
won't be changed after the first text or image is painted,
otherwise, the document will have a jarring flash which should
be avoid by most pages.

This content API will be used by embedder that are interested
in the page color scheme.

Bug:1253990

Change-Id: I5828e18eb4eb3f9190050ab53f6136c9ad5ed0a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3259232
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Commit-Queue: Michael Bai <michaelbai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949370}
NOKEYCHECK=True
GitOrigin-RevId: 19f17a30e102f811bc90a13f79e8ad39193a6403