URLPattern: Escape some suffixes trailing `:foo` groups.
This CL fixes another problem case from:

https://github.com/WICG/urlpattern/issues/145

In this case we need to detect when a group suffix could be mistaken
for trailing custom name characters.  When this happens we should escape
the first character of the suffix.  For example: `{:foo\\bar}` instead
of `{:foobar}`.

Bug: 1263673
Change-Id: I2aa2d043ef4c71433fdc0be113fcafddf5ad3532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3318605
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949707}
NOKEYCHECK=True
GitOrigin-RevId: a3bf96bb292071926a248edc7520230d461bf5bc