Record when a process exits without acking certain pagelifecycle updates.
This debugging is kind of ugly but should be temporary.

When a PageLifecycleState update which should cause a persisted pageshow
is sent, we note the timestamp. If the renderer exits before the ack
arrives, we will record the time difference in UMA and also the exit
status.

Recording the exit status is hacky because the enum does not have fixed
elements, it varies by platform.

Bug: 1234634
Change-Id: I8117a28ef4b0482a12669362b154724f59ea5be2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3321058
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949791}
NOKEYCHECK=True
GitOrigin-RevId: 6a0044b0fa21e1d4d6a9b947bf0b15d53a2c615f