Fenced Frames: Respect COEP:require-corp in the embedding frame
Fenced frames did not follow COEP configuration in an embedder
because fenced frame is under a different frame tree and it cannot
retrieve an embedder with GetParent.

In this CL, we make navigation_request to use outer document
instead of parent document to check COEP inheritance check
if the navigating frame is a root of a fenced frame tree.

Bug: 1277430
Change-Id: I4742e5bc6b526cc6ec9cf5162b40edaf56ecb03e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3316097
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949966}
NOKEYCHECK=True
GitOrigin-RevId: bd6126afcb12087a1a6843732cc3788c6f878236