Check validity of created DOMException when aborting an AbortSignal
Callers of CreateOrEmpty() are expected to check if the result is valid
or not. This is because CreateOrEmpty() may return the empty handle of
v8::Local.

Hence, this CL adds this check after creating DOMExceptions
in DOM's AbortSignal and AbortController implementation to avoid
potential null dereference errors.

Bug: 1276735
Change-Id: I1506bd4af8225f610beebb2c49912362b703a903
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3318399
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949989}
NOKEYCHECK=True
GitOrigin-RevId: 999cd2a17a5b1c70c5294ab5e4990dc2a129a429