[@layer] CSS preload scanner should preload layered imports
CSS preload scanner used to treat everything in an @import rule after
the url as the media condition. This patch makes it further parse a
'layer' keyword or 'layer()' function following the url, so that layered
imports can also be preloaded.

This patch also fixes a bug that other at-rules before an @import rule
makes it unable to preload. This is crucial to cascade layers because
having empty @layer statements before imports is an important usage
pattern.

Bug: 1095765
Change-Id: I3aa65966f065aa4fca6d896ad11f82baf4c9d8e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3321438
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950225}
NOKEYCHECK=True
GitOrigin-RevId: d53bf2d989eb0756ae2689f1b16c2875e5384e5e