[@container] Tests for querying content box
The test expectations and the current behavior in Blink is that we are
matching the content box before the scrollbar size is subtracted. It is
not entirely clear for me if that matches the specs, but it is similar
to what media queries does.

Bug: 1269196
Change-Id: Ib13eaac4409137106090bc127b6c26722024c234
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3329861
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950549}
NOKEYCHECK=True
GitOrigin-RevId: cae8eee0742bb63cf95149afefc11e1b21dd2900