ImageDataStorageFormat clean-up
Move the various "parse from string" and "convert back to string"
functions to be in graphics_types.cc, like the rest of the enums.

Simplify deserialization code to rely on the validation of width
and height that is done by ImageData::ValidateAndCreate.

Bug: 1274220
Change-Id: Id5c78d4665ab6ecb62d39bf6f48d2a757f2caee0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3315497
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950713}
NOKEYCHECK=True
GitOrigin-RevId: 0036e4325aaffaa5b92680eab02c2c5114177790