Always assume that non-auto 'columns' might create a multicol.
We can't really tell for sure, before generating the layout box tree,
whether an element is going to create a multicol container or not (which
isn't surprising, but somehow I thought it was worth a try).

This is a partial revert of CL:3320292. This part isn't really important
anymore, because CL:3323072 works around the same issue, and also
because LayoutNGBlockFragmentation is now enabled for testing.

Note that the new test would only crash if LayoutNGBlockFragmentation is
disabled (but not anymore).

Bug: 1278531
Change-Id: I80e041a076541636be7c4a7e44c0ac91805efba2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3329901
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951029}
NOKEYCHECK=True
GitOrigin-RevId: 7cd9af914b53a044255f8d4211345fcef318fda5