Make StringImpl::ref_count_ atomic
Convert the field to an atomic type and adjust accesses to match while
preserving existing safety checks. This is not the last step to make
Blink strings thread-safe, but it is the most potentially impactful on
performance. Perf testing results and discussion are in this thread:
https://groups.google.com/a/chromium.org/g/platform-architecture-dev/c/Cd483BcI3c8

Change-Id: I2c503b782f028aebbfeb733ceb8009eb00ea8210
Bug: 1083392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3152354
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#951084}
NOKEYCHECK=True
GitOrigin-RevId: c207fe32b58967816c8fe34eadf87f9f846f9089