Dictation: Add accessibility private API to update the UI bubble
This change builds on the parent change by introducing an API that
can be used to control the visibility and content of the new Dictation
UI bubble. The follow-up change will add calls to this new API from
the Dictation extension.

AX-Relnotes: N/A
Bug: 1252037
Change-Id: I69e0c9e442c8522d042a1e1d77a2962ac5c16cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3315115
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951272}
NOKEYCHECK=True
GitOrigin-RevId: f5c8c338e2c5af85afd25f6f52e370ddb4297015