Ensure slot element with display:contents are included
A performance fix landed in version 96 which determined whether or not
a slot and its subtree were "relevant" to accessibility. Irrelevant
slot subtrees stopped being included in the accessibility tree, firing
events, etc., improving performance during slot updates.

There are two instances where a slot will lack a layout object but is
still "relevant" in terms of the accessibility tree: canvas fallback
content and slots with a computed style of display:contents. We were
correctly handling the former, but failing to handle the latter.

This change checks the computed style of non-fallback-content slots
and treats them as relevant for accessibility if they have a style
with display:contents.

AX-Relnotes: Fixes regression which caused descendants inside slots with
display:contents to be pruned from the accessibility tree.

Bug: 1278446
Change-Id: I6cd255caa61d5ad58a61fc72e42eb5d0baea3109
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3338349
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#951577}
NOKEYCHECK=True
GitOrigin-RevId: 96ec3ff17c146807c8089ea049d7a9180ec7783f