Fix web-exposed First Input Delay computation for desktop.
On desktop, when we click, the input event sequence would be pointerdown
,mousedown, pointerup, mouseup. The first input should be the
pointerdown. But our current FID computation would regard the mousedown
as the first input. And the input delay for the mousedown and
pointerdown can be different. This CL is to fix the web-exposed FID by
delaying reporting the mousedown until we receive a mouseup.

Bug: 1159951
Change-Id: I04f32745558885c6200f779570b6975ca1f4663a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3308556
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Nicolás Peña <npm@chromium.org>
Commit-Queue: Hongbo Song <hbsong@google.com>
Cr-Commit-Position: refs/heads/main@{#951697}
NOKEYCHECK=True
GitOrigin-RevId: 17d9569bf7be1ef933cc2b6d7669c5015b892272