FS: Consolidate cdm_file_system_id and cdm_type into CdmType
There are currently two mostly redundant fields used to identify a
CDM type.

cd_file_system_id was initially used by the CDM running as a pepper
plugin, so the ID maps to a pepper plugin MIME type. Its value
(the MIME type) is no longer relevant. However, this is used in the
codebase to effectively map to the CdmInfo.type, a base::Token which
uniquely identifies CDMs (Widevine, ClearKey, etc).

This CL combines these fields into a CdmType which contains both a
base::Token (CdmInfo.type) and an ID used to access to Plugin Private
File System. Once we migrate CDM off of the Plugin Private File System,
this ID will no longer be needed and CdmType can be simply aliased to
a base::Token.

This CL is the first step in the migration of CDM off of the
Plugin Private File System.

Bug: 1231162
Change-Id: Ic481ffccc5d9b8dfd214b73e17cf649ddd861608
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3257089
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951789}
NOKEYCHECK=True
GitOrigin-RevId: 1fd07819d9f403c16c0ed58498900db98a52842e