[Sheriff] Undo 'standalone' for blocklisted-service-in-filter
This is a partial revert of crrev.com/c/3322017 which converted a bunch
of bluetooth requestDevice tests to "standalone window" tests. That made
this particular test flaky on Webkit Linux Leak, so this CL reverts this
one test back to the old version.

Bug: 994756, 1280218
Change-Id: Id539b93a0b47187e8124f615c3dd6897a8f1ef59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3341457
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951913}
NOKEYCHECK=True
GitOrigin-RevId: 004daf697defe5186b7ffb304eb6ee563852f0db