Fix ChromiumPerf/mac-m1_mini_2020-perf/blink_perf.css regression
Perf test regression was reported after a CL merged.
- CL commit : a3beacbe1eaa89b5e1555fea8e017df4404d0904
- Regression Bug: https://crbug.com/1279044
- TC : blink_perf.css/PseudoClassSelectors/PseudoClassSelectors.html

The regression was from the changes in the CSSSelectorParser.
- CSSSelectorParser::ConsumePseudo

To fix the performance regression, removed the changes of adding flags
to CSSSelector, CSSSelectorParser, SelectorCheckingContext.

To set the computed style flags, iterate has argument selector to check
whether it contains pseudo in the argument when matching ':has'.

This is the perf test result with this change.
- pinpoint: https://pinpoint-dot-chromeperf.appspot.com/job/106a0b0ef20000
- result: (8,194.124 -> 8,404.958 : +2.6%)

Bug: 1279044
Change-Id: If2cadf3323aac6af59bf39728d691c4f74feba25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3338229
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Byungwoo Lee <blee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#952586}
NOKEYCHECK=True
GitOrigin-RevId: bd6c8bb6f9cbfd04f2abc7de03f1edcda0e9564d