Fix a rare race causing deadlock in WebRtcTimer.
WebRtcTimer::SchedulableCallback is a callback that runs on the
metronome. The callback runs as follows: first the metronome's
ListenerHandle acquires its is_active_lock_ and then
SchedulableCallback::MaybeRun() acquires its active_lock_.

But ScheduleableCallback::Inactivate() acquires the active_lock_ first
and then removes the listener, which if done synchronously makes the
ListenerHandle try to acquire its is_active_lock_ second.

This can cause a deadlock, so Inactivate() is updated to remove the
listener in a PostTask. We already did this in some cases, but now we
do is unconditionally.

Bug: chromium:1281399
Change-Id: Ida175261e6d4a3a5ac1ab8f80950ff8c8b9fef2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3350645
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Markus Handell <handellm@google.com>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#953197}
NOKEYCHECK=True
GitOrigin-RevId: 90aa9de0c4ff030c444982e4462427ed48483920