Do not expose aria-hidden content in name from toggle buttons.
A button with aria-pressed becomes a toggle button. It takes a
different code path in AXNodeObject::TextFromDescendants because
it's not allowed to expose children in the accessibility tree, unlike
a regular button. This causes the code to resort to innerText, which
exposes aria-hidden content.

This solution allows children in our internal accessibility tree, for
nodes with native role kToggleButton. It permits the creation of a
subtree off the button, so AXNodeObject::TextFromDescendants will
traverse it, getting the expected name.

Related: https://crrev.com/c/1883619

Bug: 1280254
Change-Id: I0aab8f2c6b78872d9dc36dd16f7959d1e671c789
AX-relnotes: do not expose aria-hidden in name from toggle buttons
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3347241
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Jacobo Aragunde Pérez <jaragunde@igalia.com>
Cr-Commit-Position: refs/heads/main@{#953242}
NOKEYCHECK=True
GitOrigin-RevId: 2ecbcfc2973456cefd41596fc585bd5285a7ce60