Fix player title logging from webcodecs
Devtools displays players with |FrameTitle|PlayerTitle| now, and its
incorrect to state that the |FrameTitle| was "...(Webcodecs)". So now
it pulls the frame title either from:
Page:
 document name, or <title /> element as a backup
Worker:
 worker name, or worker js file url as a backup.

It then sends a kLoad event with the player name, which is the part
that identifies it as webcodecs.

Change-Id: I900c9ebe0d93ed9c2a890e4d8b7508237579e4ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3351033
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#953349}
NOKEYCHECK=True
GitOrigin-RevId: ba1520239dc88022cd8dab6d341f9e7f6ba1b7ef