Deflake mozilla-focus-blur.html
This test does some sort of "roundTripToBrowser" to wait for a focus
event which it expects to occur, but sometimes the focus event comes
before this wait completes, causing the test to fail.

By waiting for the focus event itself instead of "roundTripToBrowser",
the test will stop flaking. Unfortunately, this means that if the test
fails on this step it will timeout instead of having a text failure, but
I don't know if there's something better to wait for.

There is also another usage of roundTripToBrowser earlier in this test,
but that one can't be removed so easily because it expects the focus
event to not occur...

Fixed: 1273610
Change-Id: I35a36dcec3bd1022318f183bf90c9361de5a61ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3321311
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#953985}
NOKEYCHECK=True
GitOrigin-RevId: d70167e8dbca72c486faf1c762810a71d7263633