Add accept header when downloading an image from context menu
When downloading an image from context menu, the download request is
missing the accept header. This may cause the server to return an image
that is different from the one on the page.
This CL passes the same image accept header as blink uses during context
menu download.

BUG=1243412

Change-Id: I0da04829613a61f4b6ebdd5ab4a4946f9b0067b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3355474
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#954328}
NOKEYCHECK=True
GitOrigin-RevId: 7e82cc72eec68271c0f5568ceb79e9e7eb064463