Reimplement LayoutBox::GetBackgroundPaintedExtent()
Let it call LayoutBox::PhysicalBackgroundRect(kBackgroundPaintedExtent)
instead of BackgroundImageGeometry::Calculate() to allow the latter
to use GeometryMapper.

Another reason of this CL is that the previous pixel-snapped background
extent rect may be incorrect if the object has subpixel accumulation.

One drawback of this CL is that BackgroundPaintedExtent() may be
bigger for non-repeating background images, but the case seems rare.

Bug: 732934
Change-Id: I0e4a49222adb29711f8b6d961e568a06b380ac90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3361620
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#954881}
NOKEYCHECK=True
GitOrigin-RevId: 4b8c7923b347b838075663d502540b6d1380ea62