Always calculate accessible name out of accessibility tree.
AXNodeObject::TextFromDescendants was traversing the DOM tree
instead of the accessibility tree, to account for cases
that required calculation from nodes that were not part of the tree.
The main examples of this are certain labels, which were excluded from
the tree to prevent speech duplication between label + control, and
also name from a hidden subtree via aria-labelledby or describedby.
Also affects <portal> default titles and <ruby> descriptions.

After a set of patches to include any necessary nodes to the
accessibility tree, this CL enables calculating the name or
description out of the tree, instead of running a DOM traversal.
This greatly simplifies the implementation of TextFromDescendants.

This change collaterally fixes a problem with missing whitespace in
the accessible name. The cause was that the existing DOM traversal
removed all empty or whitespace text nodes, but they are present in
the cached tree as needed. Generally, the conditions for nodes to be
included in the name that were coded in TextFromDescendants are
now part of the accessible tree construction, mainly in
AXObject::ComputeAccessibilityIsIgnoredButIncludedInTree.

The existing fix for name from ::before and ::after pseudonodes
(crbug.com/1199303) was to traverse the accessibility tree in that
case, so this patch will preserve that behavior and keep it fixed.

Bug: 1190132, 1199303, 1255036, 1276936
Change-Id: Ie17b5b951992affa5e690e2eededf224c322a493
AX-relnotes: include missing whitespace in accnames.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3320230
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#955014}
NOKEYCHECK=True
GitOrigin-RevId: 4bcb100fd89d083ab4e1b98d51df41cf95f6b909